progress on copy

+++ pkg/storage/queue/models/models.go
@@ -9,6 +9,6 @@ type FileUploadMessage struct {
 type CopyDataMessage struct {
 	SourceID         int64  `json:"source_id"`
 	Query            string `json:"query"`
-	DestinationID    int64  `json:"destination_id"`
+	DestinationID    uint   `json:"destination_id"`
 	DestinationTable string `json:"destination_table"`
 }
